timer view done // problems with time left to bet until end of turn

+++ app/src/main/java/com/example/urbalog/NetworkHelper.java
@@ -264,7 +264,7 @@ public class NetworkHelper implements Serializable {
                                     timer = new CountDownTimerHandler(currentGame.getTurnDur() * 1000, new CountDownTimerHandler.TimerTickListener() {
                                         @Override
                                         public void onTick(long millisLeft) {
-                                            PlayerConnexionActivity.net.currentPlayerView.setTimeLeftTurnTimer((int)millisLeft);
+                                            PlayerConnexionActivity.net.currentPlayerView.setTimeLeftTurnTimer(millisLeft);
                                         }
 
                                         @Override
